add abort-on-property-write scriptlet, make defineProxyChain function universal
