+++ scriptlets/src/index.ts
@@ -9,3 +9,4 @@ export { jsonPruneFetchResponse } from './json-prune-fetch-response';
 export { jsonPruneXHRResponse } from './json-prune-xhr-response';
 export { abortCurrentInlineScript } from './abort-current-inline-script';
 export { abortOnPropertyRead } from './abort-on-property-read';
+export { abortOnPropertyWrite } from './abort-on-property-write';
